requirements.txt
tree_sitter
colorama
pyyaml
setuptools
